refactor(migrate): extract helper functions to migrateHelpers.ts
Extract 12 migration helper functions from the monolithic migrate.ts
(3,438 lines) into a dedicated migrateHelpers.ts module:

- removeMiniAppIconsFromState
- removeMiniAppFromState
- addMiniApp
- addProvider
- fixMissingProvider
- addOcrProvider
- updateProvider
- addWebSearchProvider
- updateWebSearchProvider
- addSelectionAction
- addShortcuts
- addPreprocessProviders

This reduces migrate.ts by ~160 lines and makes the helper functions
reusable and independently testable.
